FEATURE: use svg=true to force graphviz to avoid image conversion
Image conversion misses some of the svg features so sometimes images do not
look right.

The tradeoff here is that we get jitter on page load at the moment, but
at least we have a way of sanely generating graphs.

Also adds some CSS to make sure graphviz graphs always fit on page.