FIX: Correct graph also when newline or comment separation
When e.g. two nodes were separated only with newlines and/or comments,
only one node was generated which hade the name of two node's name
concatenated because the newlines and comments has been removed.

The solution is to not remove anything within the DOT source
code. Instead only trimming of whitespace at the beginning and end is
made since this can be considered part of the [graphviz] BBCode tags.

Fixes https://github.com/discourse/discourse-graphviz/issues/2