FEATURE: upgrade to graphviz 2.44 from 2.40
This introduces a new script (upgrade-viz-js) which can be used to upgrade
graphviz.

It uses a maintained fork of viz.js at: https://github.com/aduh95/viz.js

This new fork ships the old asm.js version of viz js which is easy to bundle
and run.

It fork also ships webassembly versions, but there are many security
implications of running webassembly in the browser including CSP and CORS.
So we avoid webassembly. (despite mini racer being able to run it)